Implement product image upload functionality and integrate Supabase for storage

+++ app/api/products/route.ts
@@ -1,5 +1,6 @@
 import { NextResponse } from "next/server";
 import prisma from "@/app/lib/prisma";
+import supabase from "@/app/lib/supabase";
 
 export async function GET() {
   const products = await prisma.product.findMany();
@@ -8,25 +9,48 @@ export async function GET() {
 }
 
 export async function POST(request: Request) {
-  const body = await request.json();
+  const formData = await request.formData();
 
-  if (!body.name || !body.description || body.price === null || !body.image) {
+  const name = formData.get("name") as string;
+  const description = formData.get("description") as string;
+  const price = Number(formData.get("price"));
+  const image = formData.get("image");
+
+  if (!(image instanceof File)) {
+    return NextResponse.json(
+      { error: "Image harus berupa file" },
+      { status: 400 },
+    );
+  }
+
+  if (!name || !description || price === null || !image) {
     return NextResponse.json({ error: "Data tidak lengkap" }, { status: 400 });
   }
 
-  if (typeof body.price !== "number" || body.price < 0) {
+  if (typeof price !== "number" || price < 0) {
     return NextResponse.json(
       { error: "Price harus berupa angka dan lebih dari 0" },
       { status: 400 },
     );
   }
 
+  const fileName = `${Date.now()}-${image.name}`;
+
+  const { error } = await supabase.storage
+    .from("product-image")
+    .upload(fileName, image);
+
+  if (error) {
+    console.log(error);
+    return NextResponse.json({ error: "Gagal upload image" }, { status: 500 });
+  }
+
   const product = await prisma.product.create({
     data: {
-      name: body.name,
-      description: body.description,
-      price: body.price,
-      image: body.image,
+      name: name,
+      description: description,
+      price: price,
+      image: fileName,
     },
   });
 
@@ -35,10 +59,24 @@ export async function POST(request: Request) {
 
 export async function DELETE(request: Request) {
   const body = await request.json();
-  const product = await prisma.product.delete({
+
+  const product = await prisma.product.findUnique({
     where: {
       id: body.id,
     },
   });
+
+  if (!product) {
+    return NextResponse.json({ error: "Id tidak ditemukan" });
+  }
+
+  await supabase.storage.from("product-image").remove([product.image!]);
+
+  await prisma.product.delete({
+    where: {
+      id: product.id,
+    },
+  });
+
   return NextResponse.json(product);
 }
